+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 ======
 
 *   (improvement) Add helper `DeferredRoute::isValidValue()`.
+*   (improvement) Add helper `DeferredTranslation::isValidValue()`.
 
 
 7.10.2